# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Textonly API",
        "title": "Textonly API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/social/": {
            "get": {
                "description": "List social data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Social"
                ],
                "summary": "List social data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "user_id",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "social_platform",
                        "name": "social_platform",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SocialListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    }
                }
            }
        },
        "/api/social/{id}": {
            "get": {
                "description": "Get social data by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Social"
                ],
                "summary": "Get social data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SocialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update social data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Social"
                ],
                "summary": "Update social data",
                "parameters": [
                    {
                        "description": "Update Social Data",
                        "name": "data.Social",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/data.Social"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.UpdateSocialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Post social data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Social"
                ],
                "summary": "Post social data",
                "parameters": [
                    {
                        "description": "Push social data",
                        "name": "SocialPostRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SocialPostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.SocialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete social data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Social"
                ],
                "summary": "Delete social data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.UpdateSocialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "data.Metadata": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "first_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "order_by": {
                    "type": "string"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "data.Social": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "link": {
                    "type": "string"
                },
                "social_platform": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "main.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
        "main.SocialListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/data.Social"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/data.Metadata"
                }
            }
        },
        "main.SocialPostRequest": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "social_platform": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "main.SocialResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/data.Social"
                },
                "metadata": {
                    "$ref": "#/definitions/data.Metadata"
                }
            }
        },
...
        "main.UpdateSocialResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "rows_affected": {
                    "type": "integer"
                }
            }
        },
...
    }
}